+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next version
+
+### ✨ Improved
+
+* Monitor LN2 e-stops during fill and before operating a valve.
+
+
 ## 0.3.10 - 2025-03-10
 
 ### ✨ Improved